requirements.txt
openai==1.14.0
requests==2.32.3
moviepy==1.0.3
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.4.0
